cruddrf/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers
from .models import State


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['url', 'username', 'email', 'groups']


class StateSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = State
        fields = ['title', 'code', 'timestamp']